리뷰이미지 공백 허용 API 수정

+++ back/src/services/reviewService.js
@@ -13,38 +13,40 @@ const reviewService = {
         console.log("여기2")
         console.log(createdNewReview)
         // console.log("created: ",createdNewReview.reviewId)
-        const originalFilename = files.reviewFile.originalFilename;
-        const extension = path.extname(originalFilename);
-        let fileName;
-
-        if(originalFilename.split(".").length > 2) {
-            const name = originalFilename.split(".");
-
-            for(let i = 0; i < name.length-1; i++) {
-                fileName += name[i];
+        if(files.reviewFile) {const originalFilename = files.reviewFile.originalFilename;
+            const extension = path.extname(originalFilename);
+            let fileName;
+            
+            if(originalFilename.split(".").length > 2) {
+                const name = originalFilename.split(".");
+    
+                for(let i = 0; i < name.length-1; i++) {
+                    fileName += name[i];
+                }
+            } else {
+                fileName = originalFilename.split(".")[0];
             }
-        } else {
-            fileName = originalFilename.split(".")[0];
+    
+            fileName = fileName + "-" + Date.now() + extension;
+    
+            const oldPath = files.reviewFile.filepath;
+            const newPath = __dirname + "/../public/reviewImages/" + fileName;
+            // const currentReviewImageInfo = await ReviewImage.findById(reviewId);
+    
+            const reviewId = createdNewReview.reviewId
+            console.log("reviewId: ", reviewId)
+            console.log("fileName: ", fileName)
+            const createdReviewImage = await ReviewImage.create(reviewId, "/public/reviewImages/" + fileName);
+            console.log("createdReviewImage: ", createdReviewImage)
+            if(!createdReviewImage) throw new Error("DB에 이미지 생성 실패");
+    
+            fs.rename(oldPath, newPath, async (err) => {
+                if(err) throw new Error("이미지 업로드 실패");
+            });
+            
+            createdNewReview.reviewImage = createdReviewImage.image
         }
-
-        fileName = fileName + "-" + Date.now() + extension;
-
-        const oldPath = files.reviewFile.filepath;
-        const newPath = __dirname + "/../public/reviewImages/" + fileName;
-        // const currentReviewImageInfo = await ReviewImage.findById(reviewId);
-
-        const reviewId = createdNewReview.reviewId
-        console.log("reviewId: ", reviewId)
-        console.log("fileName: ", fileName)
-        const createdReviewImage = await ReviewImage.create(reviewId, "/public/reviewImages/" + fileName);
-        console.log("createdReviewImage: ", createdReviewImage)
-        if(!createdReviewImage) throw new Error("DB에 이미지 생성 실패");
-
-        fs.rename(oldPath, newPath, async (err) => {
-            if(err) throw new Error("이미지 업로드 실패");
-        });
         
-        createdNewReview.reviewImage = createdReviewImage.image
         createdNewReview.errorMessage = null;
 
         console.log("location: ",createdNewReview)
@@ -109,58 +111,60 @@ const reviewService = {
             review = await Review.update({ reviewId, fieldToUpdate, newValue });
         }
 
-        const originalFilename = files.reviewFile.originalFilename;
-        const extension = path.extname(originalFilename);
-        let fileName;
-
-        if(originalFilename.split(".").length > 2) {
-            const name = originalFilename.split(".");
-
-            for(let i = 0; i < name.length-1; i++) {
-                fileName += name[i];
+        if(files.reviewFile){
+            const originalFilename = files.reviewFile.originalFilename;
+            const extension = path.extname(originalFilename);
+            let fileName;
+    
+            if(originalFilename.split(".").length > 2) {
+                const name = originalFilename.split(".");
+    
+                for(let i = 0; i < name.length-1; i++) {
+                    fileName += name[i];
+                }
+            } else {
+                fileName = originalFilename.split(".")[0];
             }
-        } else {
-            fileName = originalFilename.split(".")[0];
-        }
-
-        fileName = fileName + "-" + Date.now() + extension;
-
-        const oldPath = files.reviewFile.filepath;
-        const newPath = __dirname + "/../public/reviewImages/" + fileName;
-        const currentReviewImageInfo = await ReviewImage.findById(reviewId);
-        
-        if(!currentReviewImageInfo) {
-            // DB에 저장된 이미지가 없으면 생성
-            const createReviewImage = await ReviewImage.create(reviewId, "/public/reviewImages/" + fileName);
-
-            if(!createReviewImage) throw new Error("DB에 이미지 생성 실패");
-
-            fs.rename(oldPath, newPath, async (err) => {
-                if(err) throw new Error("이미지 업로드 실패");
-            });
-
-            review.reviewImage = createReviewImage.image
-        } else {
-            // DB에 이미가 있으면 업데이트
-            const fieldToUpdate = "image";
-            const newValue = "/public/reviewImages/" + fileName;
-            const updatedReviewImage = await ReviewImage.update(reviewId, fieldToUpdate, newValue);
+    
+            fileName = fileName + "-" + Date.now() + extension;
+    
+            const oldPath = files.reviewFile.filepath;
+            const newPath = __dirname + "/../public/reviewImages/" + fileName;
+            const currentReviewImageInfo = await ReviewImage.findById(reviewId);
             
-            review.reviewImage = updatedReviewImage.image;
-            // console.log("updatedReview: ",review)
-            fs.unlink(`src/public/reviewImages/${currentReviewImageInfo.image}`, (err) => {
-                // if(err) throw new Error("이미지 삭제 실패");
-                console.log("이미지 삭제 실패")
-            })
-
-            fs.rename(oldPath, newPath, (err) => {
-                if(err) throw new Error("이미지 업로드 실패");
-            });  
+            if(!currentReviewImageInfo) {
+                // DB에 저장된 이미지가 없으면 생성
+                const createReviewImage = await ReviewImage.create(reviewId, "/public/reviewImages/" + fileName);
+    
+                if(!createReviewImage) throw new Error("DB에 이미지 생성 실패");
+    
+                fs.rename(oldPath, newPath, async (err) => {
+                    if(err) throw new Error("이미지 업로드 실패");
+                });
+    
+                review.reviewImage = createReviewImage.image
+            } else {
+                // DB에 이미가 있으면 업데이트
+                const fieldToUpdate = "image";
+                const newValue = "/public/reviewImages/" + fileName;
+                const updatedReviewImage = await ReviewImage.update(reviewId, fieldToUpdate, newValue);
+                
+                review.reviewImage = updatedReviewImage.image;
+                // console.log("updatedReview: ",review)
+                fs.unlink(`src/public/reviewImages/${currentReviewImageInfo.image}`, (err) => {
+                    // if(err) throw new Error("이미지 삭제 실패");
+                    console.log("이미지 삭제 실패")
+                })
+    
+                fs.rename(oldPath, newPath, (err) => {
+                    if(err) throw new Error("이미지 업로드 실패");
+                });  
+            }
         }
+
+       
 
         review.errorMessage = null;
-
-
 
         return review;
     },
